We want to use Homebrew's default bottle URL
Homebrew looks to a different directory by default for finding
bottles for formulas defined in taps. For Shopify's tap, that URL
is https://homebrew.bintray.com/bottles-shopify/

Since we aren't creating the bottles ourselves, and actually want
the historic Homebrew ones, we need to explicitly specify the
default Homebrew root.