fix: stop renewing certs every minute (#745)
Based on google certificate authority logs trino certs are being requested every minute by this service.

Looks like this is because the renew_before time is set to 11h59m and the cert duration is 12h. This effectively will make certloader request a new cert every minute.

This change proposes that we extend the cert duration to 24 hours and leave the default renew_before setting of 1 hour.